Added status change callbacks.
These in form of virtual functions in the main KMotion class, to
be overriden in derived classes to receive notification of changes
in one or more fields of the MAIN_STATUS struct.  This is a more
efficient approach for (e.g.) the Python binding than having the
higher level language directly access fields in the status struct.

This requires calling Poll() regularly (similar to console handling).