Added host status notification API.
SetHostStatus() sets a 32 bit int which is used to notify the
kflop about host status (currently, only HOST_JOB_ACTIVE_BIT (=0x01) is
defined.  This is only sent on the next GetStatus request, so the
'poll' parameter (default true) is used to automatically send a
GetStatus query, by calling Poll().