New VS2013 .sln file added to manage win32 build of KMotion and GCodeInterpreter DLLs.
Python binding now builds with DLLs generated in this repo and should not require
copying any files to the Windows machine.  Because of swig.exe limitations on
Windows, we need to hard-code an absolute path for the include file -I
directories for swigwin.  So check the command line for the custom build
step.  Relies on x: mapping to \\pluto\steve-home.

To build Python binding for Win32:
Open and build BuildBasicDLLs.sln via VS2013 on Windows.
Open and build python/win/pykmotion.sln
Output files are in python/win/kmotion.py and
python/win/Release/_kmotion.pyd etc.  These need to be installed
in the appropriate location for Python to access, as well as the
basic DLLs.  TBD.