KMotionServer takes -t, -u -p args.
This was an attempt to resolve a wierd Linux bug: simplecnc starts the KMS daemon
which listens on port 57777.  simplecnc itself listens on port 9000 for WebSockets.
Close simplecnc, and the listening port gets "owned" by KMS daemon (according
to netstat) and we can't restart simplecnc.  But even though KMS now does not listen
on tcp (-u option) it still happens.  Very strange.